+++ app/javascript/vue/tasks/otu/browse/components/specimens/TypeInformation.vue
@@ -1,29 +1,42 @@
 <template>
-  <div class="panel margin-small-bottom">
+  <div class="panel margin-small-bottom padding-small">
     <div
       @click="expand = !expand"
-      class="cursor-pointer inline">
+      class="cursor-pointer inline"
+    >
       <div
         :data-icon="expand ? 'w-arrow-down' : 'w-arrow-right'"
-        class="expand-box circle-button button-default separate-right"/>
-      <span>[<span v-html="types.map(type => `${type.type_type} of ${type.original_combination}`).join('; ')"/>] <span>{{ ceLabel }}</span></span>
+        class="expand-box circle-button button-default"
+      />
+      <span class="margin-small-left">
+        [<span
+          v-html="
+            types
+              .map(
+                (type) => `${type.type_type} of ${type.original_combination}`
+              )
+              .join('; ')
+          "
+        />] - <span v-html="ceLabel" />
+      </span>
     </div>
     <template v-if="expand">
       <type-data
         class="species-information-container"
         v-for="type in types"
         :key="type.id"
-        :type="type"/>
+        :type="type"
+      />
       <specimen-information
         v-if="expand"
         class="species-information-container"
-        :specimen="specimen"/>
+        :specimen="specimen"
+      />
     </template>
   </div>
 </template>
 
 <script>
-
 import SpecimenInformation from './SpecimenInformation'
 import TypeData from './TypeData'
 
@@ -46,17 +59,28 @@ export default {
   },
 
   computed: {
-    ceLabel () {
-      const levels = ['country', 'stateProvince', 'county', 'verbatimLocality']
+    ceLabel() {
       const tmp = []
-      levels.forEach(item => {
-        if (this.specimen[item]) { tmp.push(this.specimen[item]) }
+      const levels = ['Country', 'State/Province', 'County']
+      const verbatimLabel = this.specimen.verbatimLocality
+
+      levels.forEach((level) => {
+        const levelData = this.specimen[level]
+
+        if (levelData) {
+          tmp.push(`<b>${levelData}</b>`)
+        }
       })
-      return tmp.join(', ')
+
+      if (verbatimLabel) {
+        tmp.push(verbatimLabel)
+      }
+
+      return tmp.join('; ')
     }
   },
 
-  data () {
+  data() {
     return {
       expand: false
     }
@@ -65,7 +89,7 @@ export default {
 </script>
 
 <style scoped>
-  .species-information-container {
-    margin-left: 20px;
-  }
+.species-information-container {
+  margin-left: 20px;
+}
 </style>